+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+* Adds support for P2IDR (Pay To ID with Recall) transactions on both the CLI
+  and the lib.
+
 ## 0.1.0 (2024-03-15)
 
 * Initial release.